backend/requirements.txt
fastapi
pydantic
uvicorn
pymongo
python-dotenv
beanie